feat: Update dependencies and project configuration
Refactors the project by updating various dependencies to their latest versions.
This includes changes to React, Vite, TypeScript, and related plugins.
Also, the project name in `package.json` is adjusted to `brutal-transcribe`
and the default language in `index.html` is changed to `pt-BR`.

Furthermore, `vite.config.ts` is simplified, and `tsconfig.json` is updated
to align with modern TypeScript practices and Vite's build process.
The `define` section in `vite.config.ts` is adjusted to correctly expose
`process.env`.